implementing new arch on the main front app
