--- apps/front/src/components/signup/components/generic/Loading.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-import { LoadingBlock } from '@components';
-
-import { SignupPreloadedDataLabel } from '../../types/signup-preloaded-data-label.type';
-
-type LoadingProps = {
-  label: SignupPreloadedDataLabel;
-};
-
-export const Loading = ({ label }: LoadingProps) => (
-  <LoadingBlock
-    name={`loading-${label}`}
-    text={`Loading available ${label} for you ...`}
-  />
-);
