--- apps/front/src/components/signup/components/index.ts
@@ -1,2 +1,0 @@
-export * from './Roles';
-export * from './Skills';
