--- apps/front/src/components/signup/types/query-status.type.ts
@@ -1,3 +1,0 @@
-import { QueryStatus as Status } from 'react-query';
-
-export type QueryStatus = Status | 'noData';
