+++ apps/front/src/templates/logged-user-home/hooks/useRedirectOnNoUserData.ts
