+++ apps/front/src/templates/logged-user-home/molecules/fullpage-loader/FullpageLoader.tsx
@@ -1,0 +1,9 @@
+import { CircularProgress } from '@mui/material';
+
+import { FullpageBox } from '@components/organisms';
+
+export const FullpageLoader = () => (
+  <FullpageBox>
+    <CircularProgress />
+  </FullpageBox>
+);
