+++ apps/front/src/templates/logged-user-home/molecules/index.ts
@@ -1,0 +1,4 @@
+export * from './user-avatar/UserAvatar';
+export * from './user-infos/UserInfos';
+export * from './user-skills/UserSkills';
+export * from './fullpage-loader/FullpageLoader';
