+++ apps/front/src/templates/logged-user-home/molecules/user-avatar/logic/avatar.logic.ts
