+++ apps/front/src/templates/logged-user-home/molecules/user-skills/skill-icon/logic/getRandomColor.ts
