+++ apps/front/src/templates/logged-user-home/molecules/user-skills/skill/Skill.tsx
@@ -1,0 +1,18 @@
+import { ListItem, ListItemIcon, ListItemText } from '@mui/material';
+
+import { SkillDto } from '@api';
+
+import { SkillIcon } from '../skill-icon/SkillIcon';
+
+export const Skill = ({ id, name }: SkillDto) => (
+  <ListItem key={id}>
+    <ListItemIcon
+      sx={{
+        minWidth: 40,
+      }}
+    >
+      <SkillIcon idSkill={id} />
+    </ListItemIcon>
+    <ListItemText primary={name} />
+  </ListItem>
+);
