+++ apps/front/src/templates/signup-form/organisms/index.ts
@@ -1,0 +1,2 @@
+export * from './roles/Roles';
+export * from './skills/Skills';
