+++ apps/front/src/templates/signup-form/organisms/molecules/loading-error/LoadingError.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import { ApiResponseDto } from '@api';
+import { ErrorBlock } from '@components/molecules';
+
+import { SignupPreloadedDataLabel } from '../types/signup-preloaded-data-label.type';
+
+type LoadingErrorProps = {
+  label: SignupPreloadedDataLabel;
+  error: ApiResponseDto | null;
+};
+
+export const LoadingError = ({ label, error }: LoadingErrorProps) => (
+  <ErrorBlock
+    text={error?.message ?? `An error occured while fetching ${label}`}
+  />
+);
