+++ apps/front/src/templates/signup-form/organisms/molecules/types/signup-preloaded-data-label.type.ts
